+++ api/specs/web-server/_auth.py
@@ -8,11 +8,14 @@ from typing import Any
 
 from _common import Error, Log
 from fastapi import APIRouter, status
-from models_library.api_schemas_webserver.auth import UnregisterCheck
+from models_library.api_schemas_webserver.auth import (
+    ApiKeyCreate,
+    ApiKeyGet,
+    UnregisterCheck,
+)
 from models_library.generics import Envelope
 from pydantic import BaseModel, Field, confloat
 from simcore_service_webserver._meta import API_VTAG
-from simcore_service_webserver.login.api_keys_handlers import ApiKeyCreate, ApiKeyGet
 from simcore_service_webserver.login.handlers_2fa import Resend2faBody
 from simcore_service_webserver.login.handlers_auth import (
     LoginBody,
